+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 **Fixes:**
 - Added delete idempotency to the ontap-nas-economy driver as needed by Trident.
+- Fixed an issue where qtrees with names near the 64-character limit could not
+  be deleted.
 
 **Enhancements:**